+++ editor.planx.uk/src/pages/FlowEditor/components/FeedbackLog/mocks/mockFeedback.ts
@@ -1,19 +1,207 @@
-import { Feedback } from "../../../../../routes/feedback";
+import { Feedback } from "../types";
 
 export const mockFeedback: Feedback[] = [
   {
+    address: "11, RED STREET, NEW TESTGATE, LONDON, BARNET, N11 1EE",
+    createdAt: "2025-02-26T23:23:55.181082+00:00",
+    feedbackScore: null,
+    flowName: "Find out if you need planning permission",
+    id: 994,
+    nodeTitle: "What would you like to do next?",
+    nodeType: "NextSteps",
+    type: "issue",
+    userComment: "Same error",
+    userContext: "Not loading",
+    platform: "mobile",
+    browser: "Safari",
+    helpDefinition: null,
+    helpSources: null,
+    helpText: null,
+    nodeData: {
+      tags: [],
+      steps: [
+        {
+          url: "https://www.barnet.gov.uk/planning-and-building-control/planning-applications-and-permissions?source=planx-permissionneeded",
+          title: "Get advice from our website on how to proceed ",
+          description:
+            "This link takes you to the page that is relevant to your development",
+        },
+        {
+          title: "Give us your feedback on this service",
+        },
+      ],
+      title: "What would you like to do next?",
+      description: "<p>Choose how you would like to proceed.</p>",
+    },
+    nodeId: "xyG142rLUF",
+    nodeText: "<p>Choose how you would like to proceed.</p>",
+    projectType: '["changeOfUse.let.whole"]',
+    status: "unread",
+    editorNotes: "This is important!",
+  },
+  {
     address: null,
-    createdAt: "2025-03-19T13:54:30.324494+00:00",
-    feedbackScore: 3,
-    flowName: "Feedback test 1",
-    id: 7,
-    nodeTitle: "Tell us what you think",
-    nodeType: "Feedback",
-    type: "component",
-    userComment: "Neither good nor bad",
+    createdAt: "2025-02-26T21:02:19.764282+00:00",
+    feedbackScore: null,
+    flowName: "Find out if you need planning permission",
+    id: 993,
+    nodeTitle: null,
+    nodeType: null,
+    type: "issue",
+    userComment: "Page offline",
+    userContext: "Trying to see if we need permission",
+    platform: "mobile",
+    browser: "Chrome",
+    helpDefinition: null,
+    helpSources: null,
+    helpText: null,
+    nodeData: null,
+    nodeId: null,
+    nodeText: null,
+    projectType: null,
+    status: "actioned",
+    editorNotes: "No action needed",
+  },
+  {
+    address: "4, TEST AVENUE, LONDON, SOUTHWARK, SE23 1CA",
+    createdAt: "2025-02-26T14:03:52.573405+00:00",
+    feedbackScore: null,
+    flowName: "Report a planning breach",
+    id: 992,
+    nodeTitle: "Not development",
+    nodeType: "Notice",
+    type: "issue",
+    userComment: "couldn't find a place to report the tree damage",
+    userContext:
+      "Trying to report tree pruning for various trees in the gardens of the property. The trees fall under TPO 440 and should need permission to prune or cut down",
+    platform: "desktop",
+    browser: "Safari",
+    helpDefinition: null,
+    helpSources: null,
+    helpText: null,
+    nodeData: {
+      color: "#EFEFEF",
+      title: "Not development",
+      description:
+        "<p>It looks like this project may not need planning permission because it does not meet the definition of &apos;development&apos;.</p><p>If this is the case, this would not be investigated as a breach of planning control.</p>",
+      resetButton: false,
+    },
+    nodeId: "dUn9JOO1Ne",
+    nodeText:
+      "<p>It looks like this project may not need planning permission because it does not meet the definition of &apos;development&apos;.</p><p>If this is the case, this would not be investigated as a breach of planning control.</p>",
+    projectType: '["alter.trees.tree.prune"]',
+    status: "urgent",
+    editorNotes: "Let's do something about this",
+  },
+  {
+    address: "90, TESTING LANE, LONDON, CAMDEN, NW6 8TS",
+    createdAt: "2025-02-26T11:37:03.14102+00:00",
+    feedbackScore: null,
+    flowName: "Find out if you need planning permission",
+    id: 991,
+    nodeTitle: "Tell us what we need to add",
+    nodeType: "Notice",
+    type: "unhelpful",
+    userComment:
+      "I would like to change the existing sign board at the front of the building. ",
     userContext: null,
     platform: "desktop",
     browser: "Chrome",
+    helpDefinition: null,
+    helpSources: null,
+    helpText:
+      "<p>Help us improve the service by telling what is missing from this list.</p>",
+    nodeData: {
+      info: "<p>Help us improve the service by telling what is missing from this list.</p>",
+      color: "#ffffff",
+      title: "Tell us what we need to add",
+      description:
+        "<p>If the project isn’t in the list:</p><ul><li><p>use the feedback link at the bottom of this page</p></li><li><p>we use your feedback to add to the service<br>(continue this edit)</p></li></ul>",
+      resetButton: true,
+    },
+    nodeId: "0lIzSh5fRW",
+    nodeText:
+      "<p>If the project isn’t in the list:</p><ul><li><p>use the feedback link at the bottom of this page</p></li><li><p>we use your feedback to add to the service<br>(continue this edit)</p></li></ul>",
+    projectType: '["other"]',
+    status: "in_progress",
+    editorNotes: "No action needed",
+  },
+  {
+    address: "TEST CO LTD, 122, TEST ROAD, TESTVILLE, DONCASTER, DN11 7ZY",
+    createdAt: "2025-02-25T16:19:03.647646+00:00",
+    feedbackScore: null,
+    flowName: "Find out if you need planning permission",
+    id: 990,
+    nodeTitle: "Tell us what we need to add",
+    nodeType: "Notice",
+    type: "comment",
+    userComment:
+      "I want to highlight a front step in to a dental practice. Can we do this on a building that is grade 2 listed.",
+    userContext: null,
+    platform: "desktop",
+    browser: "Microsoft Edge",
+    helpDefinition: null,
+    helpSources: null,
+    helpText:
+      "<p>Help us improve the service by telling what is missing from this list.</p>",
+    nodeData: {
+      info: "<p>Help us improve the service by telling what is missing from this list.</p>",
+      color: "#ffffff",
+      title: "Tell us what we need to add",
+      description:
+        "<p>If the project isn’t in the list:</p><ul><li><p>use the feedback link at the bottom of this page</p></li><li><p>we use your feedback to add to the service<br>(continue this edit)</p></li></ul>",
+      resetButton: true,
+    },
+    nodeId: "0lIzSh5fRW",
+    nodeText:
+      "<p>If the project isn’t in the list:</p><ul><li><p>use the feedback link at the bottom of this page</p></li><li><p>we use your feedback to add to the service<br>(continue this edit)</p></li></ul>",
+    projectType: '["other"]',
+    status: "unread",
+    editorNotes: null,
+  },
+  {
+    address: "6, TEST LANE, ST ALBANS, HERTFORDSHIRE, AL1 1AA",
+    createdAt: "2025-02-25T11:01:41.019344+00:00",
+    feedbackScore: null,
+    flowName: "Find out if you need planning permission",
+    id: 988,
+    nodeTitle: "If you can't find your project",
+    nodeType: "Notice",
+    type: "idea",
+    userComment: "add external wall insulation works",
+    userContext: null,
+    platform: "desktop",
+    browser: "Microsoft Edge",
+    helpDefinition: null,
+    helpSources: null,
+    helpText: null,
+    nodeData: {
+      color: "#EFEFEF",
+      title: "If you can't find your project",
+      description:
+        "<p>Use the feedback button at the bottom of the page to tell us about your project so that we can add it in the future.<br><br>If you didn&apos;t select any other projects, the next page shows a longer list of projects.</p>",
+      resetButton: false,
+    },
+    nodeId: "mrQ6sSQqk9",
+    nodeText:
+      "<p>Use the feedback button at the bottom of the page to tell us about your project so that we can add it in the future.<br><br>If you didn&apos;t select any other projects, the next page shows a longer list of projects.</p>",
+    projectType: null,
+    status: "actioned",
+    editorNotes: "Let's take some action here - BB",
+  },
+  {
+    address: "TEST HOUSE, 1A, TEST ROAD, LONDON, CAMDEN, NW3 8CW",
+    createdAt: "2025-02-25T08:56:57.86789+00:00",
+    feedbackScore: 3,
+    flowName: "CAMDEN - Apply for planning permission",
+    id: 986,
+    nodeTitle: "Tell us what you think",
+    nodeType: "Feedback",
+    type: "component",
+    userComment: "",
+    userContext: null,
+    platform: "desktop",
+    browser: "Microsoft Edge",
     helpDefinition: null,
     helpSources: null,
     helpText: null,
@@ -22,14 +210,14 @@ export const mockFeedback: Feedback[] = [
       disclaimer:
         "The information collected here isn't monitored by planning officers. Don't use it to give extra information about your project or submission. If you do, it cannot be used to assess your project.",
       description:
-        'This service is a work in progress, any feedback you share about your experience will help us to improve it.\n<br>\n<br>\nDon\'t share any personal or financial information in your feedback. If you do we will act according to our <a href="">privacy policy</a>',
+        'This service is a work in progress, any feedback you share about your experience will help us to improve it.\n<br>\n<br>\nDon\'t share any personal or financial information in your feedback. If you do we will act according to our <a href="">privacy policy</a>.',
       ratingQuestion:
         "<strong>How would you rate your experience with this service?</strong>",
       feedbackRequired: false,
       freeformQuestion:
         "<strong>Please tell us more about your experience.</strong>",
     },
-    nodeId: "SXap6mrSSo",
+    nodeId: "BLx7CSrWDv",
     nodeText:
       'This service is a work in progress, any feedback you share about your experience will help us to improve it.\n<br>\n<br>\nDon\'t share any personal or financial information in your feedback. If you do we will act according to our <a href="">privacy policy</a>',
     projectType: null,
