+++ editor.planx.uk/src/pages/FlowEditor/components/FeedbackLog/queries/updateEditorNotes.ts
@@ -1,6 +1,7 @@
 import gql from "graphql-tag";
 import { client } from "lib/graphql";
-import { Feedback } from "routes/feedback";
+
+import { Feedback } from "../types";
 
 export const updateEditorNotes = async (updatedRow: Feedback) => {
   await client.mutate({
